New functionality in the connect window dubbelclick a row will connect single click will move the data down, add and remove buttons will not modify the entries anymore
git-svn-id: https://wokjab.svn.sourceforge.net/svnroot/wokjab/wokjab/trunk/wokjab@145 a4058eb7-c820-0410-9b0b-94ec00e1a786

+++ wokjab/src/GUIMessageWidget.cc
@@ -319,8 +319,7 @@ GUIMessageWidget::tw1_event_after (GtkWidget *text_view, GdkEvent  *ev, GUIMessa
  if (gtk_text_iter_get_offset (&start) != gtk_text_iter_get_offset (&end))
    return FALSE;
 			
-	gtk_text_view_window_to_buffer_coords (GTK_TEXT_VIEW (text_view), GTK_TEXT_WINDOW_WIDGET,
-																																								event->x, event->y, &x, &y);
+	gtk_text_view_window_to_buffer_coords (GTK_TEXT_VIEW (text_view), GTK_TEXT_WINDOW_WIDGET,(gint)event->x, (gint)event->y, &x, &y);
 
 	gtk_text_view_get_iter_at_location (GTK_TEXT_VIEW (text_view), &iter, x, y);
